Added policy layer,
this is where validation and business rules are separated from the service classes.

Service class now independently handle databaase and routes integration, validation should no longer process here
