+++ app/policies/__init__.py
@@ -1,0 +1,3 @@
+from app.policies.p_UserPolicy import UserPolicy
+
+UserPolicy_INS = UserPolicy()
