+++ app/templates/auth/pages/income.html
@@ -7,6 +7,7 @@
     <!-- Main Content -->
     <div id="content">
 
+
         <!-- Topbar -->
             {% include 'auth/partials/topbar.html' %}
         <!-- End of Topbar -->
@@ -14,12 +15,77 @@
         <!-- Begin Page Content -->
         <div class="container-fluid">
 
+            
             <!-- Page Heading -->
             <h1 class="h3 mb-2 text-gray-800">Tables</h1>
             <p class="mb-4">DataTables is a third party plugin that is used to generate the demo table below.
                 For more information about DataTables, please visit the <a target="_blank"
                     href="https://datatables.net">official DataTables documentation</a>.</p>
 
+<!-- Button trigger modal -->
+<button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal">
+  Launch demo modal
+</button>
+
+<!-- Modal -->
+<div class="modal fade" id="exampleModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+  <div class="modal-dialog" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h5 class="modal-title" id="exampleModalLabel">Add Income</h5>
+        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+      </div>
+      <div class="modal-body">
+        
+        <form>
+            <div class="form-group">
+                <label for="input_category_name">Category name</label>
+                <input type="text" class="form-control" id="input_category_name" placeholder="Salary, Gambling win, Profit, etc...">
+            </div>
+            <div class="form-row">
+                <div class="form-group col-md-6">
+                <label for="input_source">Source</label>
+                <input type="email" class="form-control" id="input_source" placeholder="Email">
+                </div>
+                <div class="form-group col-md-6">
+                <label for="input_amount">Amount</label>
+                <input type="number" class="form-control" id="input_amount" placeholder="0.00">
+                </div>
+            </div>
+            <div class="form-row">
+                <div class="form-group col-md-6">
+                    <label for="input_received_date">Received Date</label>
+                    <input type="date" class="form-control" id="input_received_date" placeholder="Email">
+                </div>
+                <div class="form-group col-md-6">
+                    <label for="input_method">Method</label>
+                    <select class="form-control" id="input_amount">
+                        <option selected ="cash">Cash</option>
+                        <option value="gcash">G-Cash</option>
+                        <option value="bank">Bank</option>
+                        <option value="card">Card</option>
+                        <option value="other">Other</option>
+                    </select>
+                    </div>
+            </div>
+            <div class="form-group">
+                <label for="input_remarks">Remarks</label>
+                <textarea type="text" class="form-control" id="input_remarks" placeholder="Apartment, studio, or floor">
+                </textarea>
+            </div>
+            <button type="submit" class="btn btn-primary">Sign in</button>
+        </form>
+
+      </div>
+      <div class="modal-footer">
+        <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+        <button type="button" class="btn btn-primary">Save changes</button>
+      </div>
+    </div>
+  </div>
+</div>
             <!-- DataTales Example -->
             <div class="card shadow mb-4">
                 <div class="card-header py-3">
@@ -75,19 +141,4 @@
 <!-- End of Main Content -->
 
 {% endblock %}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
