+++ app/utils/exceptions/PolicyError.py
@@ -1,0 +1,2 @@
+class PolicyError(Exception):
+    pass
